Player, PlayerDTO - changed Short to Long, added Builder to DTO
TeamCategoryDTO - added

+++ demo-rest/src/main/java/co/inventorsoft/academy/demorest/entity/Player.java
@@ -22,7 +22,7 @@ public class Player {
     private Long id;
     private String playerName;
     private String nationality;
-    private Short shirtNumber;
+    private Long shirtNumber;
     private PlayerPosition playerPosition;
 
     @ManyToOne
